Separated out the functionality of the directory item view model
There are now two new classes. One for files and one for folders
