+++ bulb/source/ViewModels/FileExplorer/DirectoryItemViewModel.cs
@@ -1,4 +1,3 @@
-using System.Collections.ObjectModel;
 using System.Diagnostics;
 using System.IO;
 using System.Windows;
@@ -13,13 +12,13 @@ namespace Garlic.Bulb {
     /// <summary>
     /// Contains information for an entire single directory or file.
     /// </summary>
-    public class DirectoryItemViewModel : ViewModel {
+    public abstract class DirectoryItemViewModel : ViewModel {
         /// <summary>
         /// Name of this item.
         /// </summary>
         public string Name {
             get => name;
-            private set {
+            protected set {
                 name = value;
                 OnPropertyChanged(nameof(Name));
             }
@@ -29,14 +28,14 @@ namespace Garlic.Bulb {
         /// <summary>
         /// The Virtual File System path of this item.
         /// </summary>
-        public string VfsPath { get; private set; }
+        public string VfsPath { get; protected set; }
 
         /// <summary>
         /// Full system path of this item.
         /// </summary>
         public string FullPath => ConvertVfsPathToSystemPath(VfsPath);
 
-        public ObjectType Type { get; }
+        public abstract ObjectType Type { get; }
 
         /// <summary>
         /// Parent directory of this item.
@@ -50,62 +49,10 @@ namespace Garlic.Bulb {
         public ItemEventHandler OnOpened;
         public ItemEventHandler OnDeleted;
 
-        /// <summary>
-        /// A list of all directories within this directory. Will be empty if this item is a file.
-        /// </summary>
-        public ObservableCollection<DirectoryItemViewModel> SubDirectories { get; } = new ObservableCollection<DirectoryItemViewModel>();
-
-        /// <summary>
-        /// A list of all files within this directory. Will be empty if this item is a file.
-        /// </summary>
-        public ObservableCollection<DirectoryItemViewModel> Files { get; } = new ObservableCollection<DirectoryItemViewModel>();
-
-        /// <summary>
-        /// A list of every single item within this directory. Will be empty if this item is a file.
-        /// </summary>
-        public ObservableCollection<DirectoryItemViewModel> AllItems { get; } = new ObservableCollection<DirectoryItemViewModel>();
-
-        private readonly FileSystemWatcher watcher;
-
-        public DirectoryItemViewModel(string vfsPath) : this(new DirectoryInfo(ConvertVfsPathToSystemPath(vfsPath))) { }
-
-        public DirectoryItemViewModel(DirectoryInfo directory) : this() {
-            Name = directory.Name;
-            VfsPath = ConvertSystemPathToVfsPath(directory.FullName);
-
-            Type = ObjectType.Directory;
-
-            watcher = new FileSystemWatcher(FullPath, "*.*") {
-                EnableRaisingEvents = true
-            };
-            watcher.Created += OnFileCreated;
-            watcher.Deleted += OnFileDeleted;
-            watcher.Renamed += OnFileRenamed;
-
-            foreach (DirectoryInfo dir in directory.EnumerateDirectories()) {
-                DirectoryItemViewModel vm = CreateItem(dir);
-
-                SubDirectories.Add(vm);
-                AllItems.Add(vm);
-            }
-            foreach (FileInfo file in directory.EnumerateFiles()) {
-                DirectoryItemViewModel vm = CreateItem(file);
-
-                Files.Add(vm);
-                AllItems.Add(vm);
-            }
-        }
-
-        public DirectoryItemViewModel(FileInfo file) : this() {
-            Name = file.Name;
-            VfsPath = ConvertSystemPathToVfsPath(file.FullName);
-
-            Type = ObjectType.File;
-        }
-
-        private DirectoryItemViewModel() {
+        protected DirectoryItemViewModel(DirectoryItemViewModel parent) {
             OpenCommand = new RelayCommand(() => OnOpened?.Invoke(this));
             DeleteCommand = new RelayCommand(() => OnDeleted?.Invoke(this));
+            Parent = parent;
         }
 
         /// <summary>
@@ -122,91 +69,14 @@ namespace Garlic.Bulb {
             File.Copy(originalPath, newPath);
         }
 
-        #region Item view model events
-        private void OnItemOpened(DirectoryItemViewModel item) {
-            //Not handled here. Just pipe up
-            OnOpened?.Invoke(item);
+        public void UpdateName(string name, string vfsPath) {
+            Name = name;
+            VfsPath = vfsPath;
         }
 
-        private void OnItemDeleted(DirectoryItemViewModel item) {
-            //Delete through the OS. The file system watcher will handle the event back to us.
-            File.Delete(item.FullPath);
-        }
-        #endregion
+        protected static string ConvertVfsPathToSystemPath(string vfsPath) => ((EditorApp)Application.Current).resolveVfsPath(vfsPath);
 
-        #region File system events
-        private void OnFileCreated(object sender, FileSystemEventArgs e) {
-            Application.Current.Dispatcher.Invoke(() => {
-                DirectoryItemViewModel vm = File.GetAttributes(e.FullPath).HasFlag(FileAttributes.Directory)
-                    ? CreateItem(new DirectoryInfo(e.FullPath))
-                    : CreateItem(new FileInfo(e.FullPath));
-                Files.Add(vm);
-                AllItems.Add(vm);
-            });
-        }
-
-        private void OnFileDeleted(object sender, FileSystemEventArgs e) {
-            Application.Current.Dispatcher.Invoke(() => {
-                DirectoryItemViewModel itemVm = null;
-                foreach (var item in AllItems) {
-                    if (item.Name == e.Name) {
-                        itemVm = item;
-                        break;
-                    }
-                }
-
-                if (itemVm != null) {
-                    DeleteItem(itemVm);
-                }
-            });
-        }
-
-        private void OnFileRenamed(object sender, RenamedEventArgs e) {
-            Application.Current.Dispatcher.Invoke(() => {
-                DirectoryItemViewModel itemVm = null;
-                foreach (var item in AllItems) {
-                    if (item.Name == e.OldName) {
-                        itemVm = item;
-                        break;
-                    }
-                }
-
-                itemVm.Name = e.Name;
-                itemVm.VfsPath = ConvertSystemPathToVfsPath(e.FullPath);
-            });
-        }
-        #endregion
-
-        private DirectoryItemViewModel CreateItem(DirectoryInfo directory) {
-            var vm = new DirectoryItemViewModel(directory);
-            vm.OnOpened += (DirectoryItemViewModel item) => OnItemOpened(item);
-            vm.OnDeleted += (DirectoryItemViewModel item) => OnItemDeleted(item);
-            vm.Parent = this;
-
-            return vm;
-        }
-
-        private DirectoryItemViewModel CreateItem(FileInfo file) {
-            var vm = new DirectoryItemViewModel(file);
-            vm.OnOpened += (DirectoryItemViewModel item) => OnItemOpened(item);
-            vm.OnDeleted += (DirectoryItemViewModel item) => OnItemDeleted(item);
-            vm.Parent = this;
-
-            return vm;
-        }
-
-        private void DeleteItem(DirectoryItemViewModel item) {
-            AllItems.Remove(item);
-            if (item.Type == ObjectType.Directory) {
-                SubDirectories.Remove(item);
-            } else {
-                Files.Remove(item);
-            }
-        }
-
-        private static string ConvertVfsPathToSystemPath(string vfsPath) => ((EditorApp)Application.Current).resolveVfsPath(vfsPath);
-
-        private static string ConvertSystemPathToVfsPath(string systemPath) {
+        protected static string ConvertSystemPathToVfsPath(string systemPath) {
             string rootSystem = ConvertVfsPathToSystemPath(".");
             return systemPath.Remove(0, rootSystem.Length).Replace("\\", "/").Insert(0, ".");
         }
