+++ bulb/source/ViewModels/FileExplorer/FileViewModel.cs
@@ -1,0 +1,13 @@
+using System.IO;
+
+namespace Garlic.Bulb {
+    public class FileViewModel : DirectoryItemViewModel {
+        public override ObjectType Type => ObjectType.File;
+
+        public FileViewModel(FileInfo file, DirectoryItemViewModel parent)
+            : base(parent) {
+            Name = file.Name;
+            VfsPath = ConvertSystemPathToVfsPath(file.FullName);
+        }
+    }
+}
